+++ backend/src/main/java/com/gradesave/backend/dto/grade/GradeDto.java
@@ -1,5 +1,6 @@
 package com.gradesave.backend.dto.grade;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -10,6 +11,6 @@ import java.util.UUID;
  *
  **/
 
-public record GradeDto(UUID performanceId, UUID projectSubjectId, Double grade) {
+public record GradeDto(UUID performanceId, UUID projectSubjectId, BigDecimal grade) {
 
 }
